requirements.txt
# HAK_GAL Suite Python Requirements
# Generated: 2025-08-23
# Python Version: 3.11+

# Core Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.0
eventlet==0.33.3

# Machine Learning & AI
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# LLM Providers
ollama>=0.1.0
openai>=1.0.0
google-generativeai>=0.3.0

# Async & Networking
httpx>=0.25.0
aiohttp>=3.9.0
websocket-client>=1.6.0

# Database
# sqlite3 is built-in

# Utilities
python-dotenv>=1.0.0
pathlib>=1.0.0
pydantic>=2.0.0

# Monitoring (optional)
psutil>=5.9.0

# Development Tools (optional)
pytest>=7.4.0
black>=23.0.0
pylint>=3.0.0

# MCP Dependencies
mcp>=0.1.0  # if available

# Additional for HAK/GAL
jsonlines>=3.1.0
regex>=2023.0.0

# Sentry Integration (optional)
sentry-sdk>=2.0.0
z3-solver
